Remove specific cases/tests to support Go version below 1.15
* go1.14 (released 2020/02/25) - new binder time.duration errors message tests
* go1.13 (released 2019/09/03) - timeout middleware, HTTPError unwrapping related
* go1.12 (released 2019/02/25) - csrf middleware related
* go1.11 (released 2018/08/24) - proxy middleware related
