--- middleware/csrf_samesite.go
@@ -1,12 +1,0 @@
-// +build go1.13
-
-package middleware
-
-import (
-	"net/http"
-)
-
-const (
-	// SameSiteNoneMode required to be redefined for Go 1.12 support (see #1524)
-	SameSiteNoneMode http.SameSite = http.SameSiteNoneMode
-)
